Переносит адаптеры из OrdersApiService в модуль orders
